Bug 1662033 - Remove localId property from mailing lists. r=mkmelin
This changes the URI of a list. I don't believe we store list URIs anywhere so that shouldn't matter.

Differential Revision: https://phabricator.services.mozilla.com/D88754

--HG--
extra : rebase_source : 087b98af4849b6e94aa6fdc053c1f255ad009389

+++ mailnews/addrbook/jsaddrbook/AddrBookManager.jsm
@@ -209,7 +209,7 @@ AddrBookManager.prototype = {
     }
     let [, scheme, fileName, tail] = uriParts;
     if (tail && types.includes(scheme)) {
-      if (scheme == "jsaddrbook" && tail.startsWith("/MailList")) {
+      if (scheme == "jsaddrbook" && tail.startsWith("/")) {
         let parent = this.getDirectory(`${scheme}://${fileName}`);
         for (let list of parent.childNodes) {
           list.QueryInterface(Ci.nsIAbDirectory);
@@ -389,14 +389,12 @@ AddrBookManager.prototype = {
     }
 
     let [, scheme, fileName, tail] = uriParts;
-    if (tail) {
-      if (tail.startsWith("/MailList")) {
-        let dir = store.get(`${scheme}://${fileName}`);
-        let list = this.getDirectory(uri);
-        if (dir && list) {
-          dir.deleteDirectory(list);
-          return;
-        }
+    if (tail && tail.startsWith("/")) {
+      let dir = store.get(`${scheme}://${fileName}`);
+      let list = this.getDirectory(uri);
+      if (dir && list) {
+        dir.deleteDirectory(list);
+        return;
       }
     }
